fetch hooks list data server side
