+++ packages/nextjs/hooks/balancer/useSwap.ts
@@ -25,7 +25,6 @@ export const useSwap = (pool: Pool, swapConfig: SwapConfig): UseSwap => {
         swapKind: swapConfig.swapKind,
         paths: [
           {
-            protocolVersion: 3 as 1 | 2 | 3,
             pools: [pool.address as `0x${string}`],
             tokens: [
               {
@@ -37,6 +36,7 @@ export const useSwap = (pool: Pool, swapConfig: SwapConfig): UseSwap => {
                 decimals: tokenOut.decimals,
               }, // tokenOut
             ],
+            protocolVersion: 3 as const,
             inputAmountRaw: swapConfig.tokenIn.rawAmount,
             outputAmountRaw: swapConfig.tokenOut.rawAmount,
           },
